Extracted map method to their own class
